// class-room-sauvegarde/10/class-room-frontend/src/app/api/model/eleveDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ClasseEleveDto } from "./classeEleveDto";


export interface EleveDto { 
    idEleve?: number;
    code?: string;
    num?: number;
    nom?: string;
    dateNaissance?: string;
    gender?: EleveDto.GenderEnum;
    photo?: string;
    isAbsent?: boolean;
    selectedDate?: string; // Ajouter selectedDate
    selectedTime?: string; // Ajouter selectedTime
}
export namespace EleveDto {
    export type GenderEnum = 'MALE' | 'FEMALE';
    export const GenderEnum = {
        Male: 'MALE' as GenderEnum,
        Female: 'FEMALE' as GenderEnum
    };
}